% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_faers_structure.R
\name{check_faers_structure}
\alias{check_faers_structure}
\title{Check the structure of the FAERS folder}
\usage{
check_faers_structure(path, faersyears = years_from_faers_html())
}
\arguments{
\item{path}{(chr) The path of the folder containing a subfolder (named
"faers_raw_data") with FAERS data inside, sorted by year and quarter.}

\item{faersyears}{(chr) A vector of characters with the list of the years of
FAERS data (defaults: all years available).}
}
\value{
(lgl) TRUE (invisible) if the "faers_raw_data" exists and it is
structured correctly, FALSE (invisible) otherwise.
}
\description{
Check if exists a folder containing the FAERS data and if the structure of
the folder is correct.
}
\examples{
check_faers_structure(".")
check_faers_structure("C:/Users/pietr/Documents/Ricerca/faers")
}
