% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_year.R
\name{check_year}
\alias{check_year}
\title{Check FAERS year}
\usage{
check_year(year)
}
\arguments{
\item{year}{(int) The year of the data to download (no default)}
}
\value{
If the year is fine returns TRUE, stops
(with an error message) otherwise
}
\description{
This function check if the year selected for FAERS download is fine
}
\examples{
check_year(2018)
}
