% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose_faers_link.R
\name{compose_faers_link}
\alias{compose_faers_link}
\title{Compose FAERS link}
\usage{
compose_faers_link(
  year,
  quarter = c("q1", "q2", "q3", "q4"),
  type = c("ascii", "xml")
)
}
\arguments{
\item{year}{(int) The year of the data to download.}

\item{quarter}{(char, default: "q1") The quarter of the data
to download between "q1", "q2", "q3" or "q4".}

\item{type}{(char, default: "ascii") The format of the data to
download between "ascii" or "xml",}
}
\value{
(chr) the link to download selected FAERS \code{.zip} data.
}
\description{
This function compose the download link for FAERS data
}
\examples{
compose_faers_link(2018, "q1", "ascii")
}
