% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose_faers_path.R
\name{compose_faers_path}
\alias{compose_faers_path}
\title{Compose FAERS path}
\usage{
compose_faers_path(
  path = getwd(),
  year,
  quarter = c("q1", "q2", "q3", "q4"),
  type = c("ascii", "xml")
)
}
\arguments{
\item{path}{(char) The path of the folder where the database
will be saved (default: working directory)}

\item{year}{(int) The year of the data to download (no default)}

\item{quarter}{(char) The quarter of the data to download
("q1", "q2", "q3" or "q4", default: "q1)}

\item{type}{(char) The format of the data to download,
("ascii" or "xml", default: "ascii")}
}
\value{
Return the path of the internal folder where FAERS
data will be downloaded
}
\description{
This function compose the path where FAERS data will be downloaded
}
\examples{
compose_faers_path(year = 2020)
}
