% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_faers_folders.R
\name{create_faers_folders}
\alias{create_faers_folders}
\title{Create FAERS folder}
\usage{
create_faers_folders(
  path = getwd(),
  year,
  quarter = c("q1", "q2", "q3", "q4"),
  type = c("ascii", "xml")
)
}
\arguments{
\item{path}{(char) The path of the folder where the database
will be saved (default: working directory)}

\item{year}{(int) The year of the data to download (no default)}

\item{quarter}{(char) The quarter of the data to download
("q1", "q2", "q3" or "q4", default: "q1)}

\item{type}{(char) The format of the data to download,
("ascii" or "xml", default: "ascii")}
}
\value{
Creates a set of nested folders
}
\description{
This function creates the set of nested folders where FAERS
data will be downloaded; if the folders already exist, no nothing
}
\examples{
create_faers_folders(year = 2018, quarter = "q1", type = "ascii")
}
