% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faers.db-package.R
\docType{package}
\name{faers.db-package}
\alias{faers.db}
\alias{faers.db-package}
\title{faers.db: FAERS database in R}
\description{
Provide an R interface to retrieve, merge, and tidy
    the FAERS database.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/UBESP-DCTV/faers.db}
  \item Report bugs at \url{https://github.com/UBESP-DCTV/faers.db/issues}
}

}
\author{
\strong{Maintainer}: Corrado Lanera \email{corrado.lanera@unipd.it} (\href{https://orcid.org/0000-0002-0520-7428}{ORCID})

Authors:
\itemize{
  \item Pietro Belloni \email{pietro.belloni.1@phd.unipd.it}
  \item Nicola Guidone \email{nicola.guidone@unipd.it}
}

Other contributors:
\itemize{
  \item Giulia Lorenzoni \email{giulia.lorenzoni@unipd.it} [contributor]
  \item Dario Gregori \email{dario.gregori@unipd.it} [copyright holder]
  \item UBESP [copyright holder]
}

}
\keyword{internal}
