% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_faers_meta.R
\name{fetch_faers_meta}
\alias{fetch_faers_meta}
\title{List of FAERS data}
\usage{
fetch_faers_meta()
}
\value{
a \link[tibble:tibble-package]{tibble} reporting years, period,
quarter, and dimension for available FAERS data.
}
\description{
The function lists the metadata for the FAERS databases currently
available to download.
}
\examples{
fetch_faers_meta()
}
