% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_local.R
\name{fetch_local}
\alias{fetch_local}
\title{List of local FAERS data}
\usage{
fetch_local(path)
}
\arguments{
\item{path}{(chr) The path of the folder containing a subfolder (named
"faers_raw_data") with FAERS data inside, sorted by year and quarter.}
}
\value{
A \link[tibble:tibble-package]{tibble} reporting years, path and quarter
for local FAERS data.
}
\description{
The function lists the metadata for the FAERS databases currently in the path
chosen folder.
}
\examples{
fetch_local(".")
}
