% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_NULL_NA.R
\name{\%NULL\%}
\alias{\%NULL\%}
\title{Default for NULL (or NA) values}
\usage{
x \%NULL\% y
}
\arguments{
\item{x}{Parameter}

\item{y}{Chosen default value}
}
\value{
The parameter or the default value
}
\description{
These two functions assign a default value to a parameter if it was
previously set to NULL (or NA)
}
\examples{
NULL \%NULL\% 3 # = 3
5 \%NULL\% 3 # = 5
}
