% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_faers_html.R
\name{import_faers_html}
\alias{import_faers_html}
\title{Import FAERS html}
\usage{
import_faers_html(
  faers_url = "https://fis.fda.gov/extensions/FPD-QDE-FAERS/FPD-QDE-FAERS.html"
)
}
\arguments{
\item{faers_url}{(char) Url of the FAERS repository web page
(do not modify the default url unless it is changed by the web page owner)}
}
\value{
An XML document
}
\description{
Import FAERS html from the FAERS web page
}
\examples{
import_faers_html()
}
