% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_year_quarter_available.R
\name{is_year_quarter_available}
\alias{is_year_quarter_available}
\title{Check FAERS year}
\usage{
is_year_quarter_available(.year, .quarter)
}
\arguments{
\item{.year}{(num) year for which data are required}

\item{.quarter}{(chr) quarter for which data are required, between "q1",
"q2", "q3", "q4".}
}
\value{
(lgl) FAERS can have data for the \code{.year} and \code{.quarter} required?
}
\description{
This function check if data for the year and quarter selected are
available at FAERS to be downloaded.
}
\examples{
is_year_quarter_available(2013, "q1")  # TRUE
is_year_quarter_available(2012, "q1")  # FALSE
}
