% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_of_faers_data.R
\name{list_of_faers_data}
\alias{list_of_faers_data}
\title{List of FAERS data}
\usage{
list_of_faers_data(faers_url = NULL)
}
\arguments{
\item{faers_url}{(char) Url of the FAERS repository web page
(do not modify the default url unless it is changed by the web page
owner)}
}
\value{
a \link[tibble:tibble-package]{tibble} reporting years, period,
quarter, and dimension for available FAERS data.
}
\description{
The function lists the databases currently online on the FAERS web
page
}
\examples{
list_of_faers_data()
}
