% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_of_faers_years.R
\name{list_of_faers_years}
\alias{list_of_faers_years}
\title{Title}
\usage{
list_of_faers_years(
  faers_url = "https://fis.fda.gov/extensions/FPD-QDE-FAERS/FPD-QDE-FAERS.html"
)
}
\arguments{
\item{faers_url}{(char) Url of the FAERS repository web page
(do not modify the default url unless it is changed by the web page owner)}
}
\value{
A vector with the list of years currently online on the FAERS website
}
\description{
Title
}
\examples{
list_of_faers_years()
}
