% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_of_faers_years.R
\name{number_of_faers_years}
\alias{number_of_faers_years}
\title{Number of FAERS years}
\usage{
number_of_faers_years(faers_html)
}
\arguments{
\item{faers_html}{An output of the fetch_faers_meta() function}
}
\value{
An integer number
}
\description{
Retrieve the number of FAERS year currently listed on the FAERS web page
}
\examples{
number_of_faers_years(fetch_faers_meta())
}
