% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_faers.R
\name{read}
\alias{read}
\alias{read_demo}
\alias{read_drug}
\alias{read_indi}
\alias{read_outc}
\alias{read_reac}
\alias{read_rpsr}
\alias{read_ther}
\title{Read faers data}
\usage{
read_demo(path)

read_drug(path)

read_indi(path)

read_outc(path)

read_reac(path)

read_rpsr(path)

read_ther(path)
}
\arguments{
\item{path}{(chr) path to the FAERS \code{.txt} to read.}
}
\value{
a \link[tibble:tibble-package]{tibble} for the FAERS table in \code{path}.
}
\description{
The \code{read()} functions read FAERS data into a well formatted
tibble.
}
\section{Functions}{
\itemize{
\item \code{read_demo}: read DEMO db

\item \code{read_drug}: read DRUG db

\item \code{read_indi}: read INDI db

\item \code{read_outc}: read OUTC db

\item \code{read_reac}: read REAC db

\item \code{read_rpsr}: read RPSR db

\item \code{read_ther}: read THER db
}}

\examples{
  \dontrun{
    demo_db <- read_demo(
    path = "path_demo.csv"
    )
  }
}
