% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_faers.R
\name{read_demo}
\alias{read_demo}
\title{Title}
\usage{
read_demo(path)
}
\arguments{
\item{path}{refers to the file's path}
}
\value{
import DB read_*
}
\description{
Title
}
\examples{
read_demo <- function(path) {
 utils::read.csv(path, sep = "$")
   }
}
