% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_faers.R
\name{read_faers}
\alias{read_faers}
\title{Read faers data}
\usage{
read_faers(path)
}
\arguments{
\item{path}{(chr) path to the FAERS \code{.txt} to read.}
}
\value{
a \link[tibble:tibble-package]{tibble} for the FAERS table in \code{path}.
}
\description{
The \code{read_faers()} functions read FAERS data into a well formatted
tibble.
}
\details{
The \code{read_faers()} function can automaticaly read seven different
FAERS dataset: demo, drug, indi, outc, reac, rpsr and ther.
It convert and adapt dataset before version 2014Q4 with the new version.
Moreover, it adds a new coloums called 'period', in order to identify
the time release of dataset
}
\examples{
  \dontrun{
    faers_db <- read_faers(
    path = "path_demo21Q1.txt"
    )

    demo_db <- read_demo(
    path = "path_demo21Q1.txt"
    )

    identical(faers_db, demo_db)
    #TRUE
  }

}
