% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_qde.R
\name{retrieve_qde}
\alias{retrieve_qde}
\title{Retrieve qde}
\usage{
retrieve_qde(
  path,
  year,
  quarter = c("q1", "q2", "q3", "q4"),
  type = c("ascii", "xml"),
  create_folder = NULL,
  download_data = NULL
)
}
\arguments{
\item{path}{(chr) Path of the directory where you want to download the data}

\item{year}{(chr) The year of the data to download}

\item{quarter}{(chr) The quarter of the data to download
("q1", "q2", "q3" or "q4", default: "q1)}

\item{type}{(chr) The format of the data to download,
("ascii" or "xml", default: "ascii")}
}
\value{
(lgl) TRUE (invisible) if the download was succesful, FALSE otherwise
}
\description{
This function download the FAERS data for a specific year and quarter
}
\examples{
\dontrun{
  retrieve_qde(year = "2018", quarter = "q4")
}
a <- retrieve_qde(year="2016", quarter = "q4", path = "C:/Users/pietr/Documents/Downloads")

}
