% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_local.R
\name{update_local}
\alias{update_local}
\title{Update local FAERS data (ascii files only)}
\usage{
update_local(
  path,
  missing_metadata = what_is_missing(path),
  permission_to_update = NULL
)
}
\arguments{
\item{path}{(chr) The path of the folder containing a subfolder (named
"faers_raw_data") with FAERS data inside, sorted by year and quarter.}

\item{missing_metadata}{A \link[tibble:tibble-package]{tibble} reporting year,
quarter and type of FAERS dataset currently missing in the selected folder,
usually an output of what_is_missing() function
(default: what_is_missing(path)).}

\item{permission_to_update}{(lgl) Only if session is not in interactive mode.
TRUE: permission to download files, FALSE: deny permission to download files.}
}
\value{
(lgl) TRUE if the download was successful, FALSE otherwise.
}
\description{
This function downloads the FAERS ascii data missing in the selected folder.
}
\examples{
\dontrun{
 update_local(".")
}
}
