% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/what_is_missing.R
\name{what_is_missing}
\alias{what_is_missing}
\title{What is missing in my FAERS folder?}
\usage{
what_is_missing(path, faers_meta = fetch_faers_meta())
}
\arguments{
\item{path}{(chr) The path of the folder containing a subfolder (named
"faers_raw_data") with FAERS data inside, sorted by year and quarter.}

\item{faers_meta}{A \link[tibble:tibble-package]{tibble} with the metadata of
of FAERS data currently online (default: all the FAERS data).}
}
\value{
A \link[tibble:tibble-package]{tibble} reporting year, quarter and type
of FAERS dataset currently missing in the selected folder (invisible).
}
\description{
This function tells what FAERS datasets are not currently in the selected
folder.
}
\examples{
what_is_missing(".")
}
