% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_year.R
\name{check_year}
\alias{check_year}
\title{Check FAERS year}
\usage{
check_year(year)
}
\arguments{
\item{year}{(int) The year of the data to download.}
}
\value{
(lgl) FAERS can have data for the \code{year} required?
}
\description{
This function check if the year selected for FAERS download can have
data.
}
\examples{
check_year(2013)  # TRUE
check_year(2012)  # FALSE
}
