% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_faers.R
\name{read_demo}
\alias{read_demo}
\title{Title}
\usage{
read_demo(path)
}
\arguments{
\item{path}{refers to the dataset's path,}
}
\value{
import DB read_* already trasformed
}
\description{
Title
}
\examples{
x <- read_demo(insert/path)
}
