% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_faers.R
\name{read_demo}
\alias{read_demo}
\title{Read faers data}
\usage{
read_demo(path)
}
\arguments{
\item{path}{(chr) path to the FAERS \code{.txt} to read.}
}
\value{
a \link[tibble:tibble-package]{tibble} for the FAERS table in \code{path}.
}
\description{
The \verb{read_*()} functions read FAERS data into a well formatted
tibble.
}
\examples{
  \dontrun{
      # require q3-2020 FAERS raw data in the working directory
      read_demo("DEMO20Q3.txt")
      read_drug("DRUG20Q3.txt")
  }
}
