% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_faers.R
\name{read_demo}
\alias{read_demo}
\title{Read faers data}
\usage{
read_demo(path)
}
\arguments{
\item{path}{(chr) path to the FAERS \code{.txt} to read.}
}
\value{
a \link[tibble:tibble-package]{tibble} for the FAERS table in \code{path}.
}
\description{
The \verb{read_*()} functions read FAERS data into a well formatted
tibble.
}
\examples{
  \dontrun{
    demo_path <- system.file(
      "testing-data/DEMO20Q3-10.txt",
       package = "faers.db"
    )
    read_demo(demo_path)
  }
}
