% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_qde.R
\name{retrieve_qde}
\alias{retrieve_qde}
\title{Retrieve qde}
\usage{
retrieve_qde(
  year = format(Sys.Date(), "\%Y"),
  quarter = c("q1", "q2", "q3", "q4"),
  type = c("ascii", "xml")
)
}
\arguments{
\item{year}{The year of the data to download}

\item{quarter}{The quarter of the data to download}

\item{type}{The format of the data to download, ascii or xml}
}
\value{
The function download the selected data in the working directory
}
\description{
This function download the FAERS data for a specific year and quarter
}
\examples{
retrieve_qde(year = "2018", quarter = "q1", type = "ascii")

}
