% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_qde.R
\name{retrieve_qde}
\alias{retrieve_qde}
\title{Download FAERS data}
\usage{
retrieve_qde(
  path,
  year,
  quarter = c("q1", "q2", "q3", "q4"),
  type = c("ascii", "xml"),
  interactive_session = rlang::is_interactive(),
  create_folder = NULL,
  download_data = NULL
)
}
\arguments{
\item{path}{(chr) Path of the directory where you want to download the data}

\item{year}{(chr) The year of the data to download}

\item{quarter}{(chr) The quarter of the data to download
("q1", "q2", "q3" or "q4", default: "q1)}

\item{type}{(chr) The format of the data to download,
("ascii" or "xml", default: "ascii")}

\item{interactive_session}{(lgl) Is R running in an interactive session?
(default: the status of the current session).}

\item{create_folder}{(lgl) Only if session is not in interactive mode.
TRUE: permission to create folders, FALSE: deny permission to create folders}

\item{download_data}{(lgl) Only if session is not in interactive mode.
TRUE: permission to download data, FALSE: deny permission to download data.}
}
\value{
(lgl) TRUE if the download was successful, FALSE otherwise.
}
\description{
This function downloads the FAERS data for a specific year and quarter.
}
\examples{
\dontrun{
  retrieve_qde(path = ".", year = "2018", quarter = "q4")
}
}
