% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_qde.R
\name{retrieve_qde}
\alias{retrieve_qde}
\title{Retrieve qde}
\usage{
retrieve_qde(
  path = getwd(),
  year,
  quarter = c("q1", "q2", "q3", "q4"),
  type = c("ascii", "xml")
)
}
\arguments{
\item{path}{(char) The path of the folder where the database
will be saved (default: working directory)}

\item{year}{(int) The year of the data to download (no default)}

\item{quarter}{(char) The quarter of the data to download
("q1", "q2", "q3" or "q4", default: "q1)}

\item{type}{(char) The format of the data to download,
("ascii" or "xml", default: "ascii")}
}
\value{
Download the selected FAERS data in the chosen folder
}
\description{
This function download the FAERS data for a specific year and quarter
}
\examples{
\dontrun{
  retrieve_qde(year = 2018, quarter = "q4")
}

}
