% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_qde.R
\name{retrieve_qde}
\alias{retrieve_qde}
\title{Retrieve qde}
\usage{
retrieve_qde(
  year,
  quarter = c("q1", "q2", "q3", "q4"),
  type = c("ascii", "xml"),
  path = getwd(),
  create.folder = NULL,
  download.data = NULL
)
}
\arguments{
\item{year}{(chr) The year of the data to download (no default)}

\item{quarter}{(chr) The quarter of the data to download
("q1", "q2", "q3" or "q4", default: "q1)}

\item{type}{(chr) The format of the data to download,
("ascii" or "xml", default: "ascii")}

\item{path}{(chr) Path of the directory where you want to download the data
(default: working directory)}
}
\value{
Download the selected FAERS data in the chosen folder
}
\description{
This function download the FAERS data for a specific year and quarter
}
\examples{
\dontrun{
  retrieve_qde(year = 2018, quarter = "q4")
}

}
